Move cache logic into base
