+++ planetmapper/body_xy.py
@@ -1,19 +1,15 @@
 import datetime
-import functools
 import io
 import math
 import warnings
 from typing import (
     Any,
     Callable,
-    Concatenate,
     Iterable,
     Literal,
     NamedTuple,
-    ParamSpec,
     Protocol,
     TypedDict,
-    TypeVar,
     cast,
 )
 
@@ -32,92 +28,9 @@ from matplotlib.axes import Axes
 from matplotlib.collections import QuadMesh
 from spiceypy.utils.exceptions import NotFoundError
 
-from .body import (
-    DEFAULT_WIREFRAME_FORMATTING,
-    Body,
-    _WireframeComponent,
-    _WireframeKwargs,
-)
+from .base import _cache_clearable_result, _cache_stable_result
+from .body import Body, _WireframeComponent, _WireframeKwargs
 from .progress import progress_decorator
-
-T = TypeVar('T')
-S = TypeVar('S')
-P = ParamSpec('P')
-
-
-def _cache_clearable_result(
-    fn: Callable[Concatenate[S, P], T]
-) -> Callable[Concatenate[S, P], T]:
-    """
-    Decorator to cache the output of a method call with variable arguments.
-
-    This requires that the class has a `self._cache` dict which can be used to store
-    the cached result. The dictionary key is derived from the name of the decorated
-    function.
-
-    The results cached by this decorator can be cleared using `self._cache.clear()`, so
-    this is useful for results which need to be invalidated (i.e. backplane images
-    which are invalidated the moment the disc params are changed). If the result is
-    stable (i.e. backplane maps) then use `_cache_stable_result` instead.
-
-    Note that any numpy arguments will be converted to (nested) tuples.
-    """
-
-    @functools.wraps(fn)
-    def decorated(self, *args_in: P.args, **kwargs_in: P.kwargs):
-        args, kwargs = _replace_np_arrr_args_with_tuples(args_in, kwargs_in)
-        k = (fn.__name__, args, frozenset(kwargs.items()))
-        if k not in self._cache:
-            self._cache[k] = fn(self, *args, **kwargs)  # type: ignore
-        return self._cache[k]
-
-    return decorated
-
-
-def _cache_stable_result(
-    fn: Callable[Concatenate[S, P], T]
-) -> Callable[Concatenate[S, P], T]:
-    """
-    Decorator to cache stable result
-
-    Very roughly, this is a type-hinted version of `functools.lru_cache` that doesn't
-    cache self.
-
-    See _cache_clearable_result for more details.
-    """
-
-    @functools.wraps(fn)
-    def decorated(self, *args_in: P.args, **kwargs_in: P.kwargs):
-        args, kwargs = _replace_np_arrr_args_with_tuples(args_in, kwargs_in)
-        k = (fn.__name__, args, frozenset(kwargs.items()))
-        if k not in self._stable_cache:
-            self._stable_cache[k] = fn(self, *args, **kwargs)  # type: ignore
-        return self._stable_cache[k]
-
-    return decorated
-
-
-def _replace_np_arrr_args_with_tuples(args, kwargs) -> tuple[tuple, dict[str, Any]]:
-    args = tuple(_maybe_np_arr_to_tuple(a) for a in args)
-    kwargs = {k: _maybe_np_arr_to_tuple(v) for k, v in kwargs.items()}
-    return args, kwargs
-
-
-def _maybe_np_arr_to_tuple(o: Any) -> Any:
-    if isinstance(o, np.ndarray):
-        return _to_tuple(o)
-    return o
-
-
-def _to_tuple(arr: np.ndarray):
-    if arr.ndim > 1:
-        return tuple(_to_tuple(a) for a in arr)
-    elif arr.ndim == 1:
-        return tuple(arr)
-    elif arr.ndim == 0:
-        return float(arr)
-    else:
-        raise ValueError(f'Error converting arr {arr!r} to tuple')
 
 
 class _MapKwargs(TypedDict, total=False):
@@ -308,9 +221,6 @@ class BodyXY(Body):
         """
 
         # Run setup
-        self._cache = {}
-        self._stable_cache = {}
-
         self._nx: int = nx
         self._ny: int = ny
 
@@ -347,14 +257,6 @@ class BodyXY(Body):
             self._rotation_radians,
             super()._get_equality_tuple(),
         )
-
-    # Cache management
-    def _clear_cache(self):
-        """
-        Clear cached results from `_cache_result`.
-        """
-        # TODO document cache clearing (incl stable cache)
-        self._cache.clear()
 
     # Coordinate transformations
     @_cache_clearable_result
@@ -1458,9 +1360,9 @@ class BodyXY(Body):
             )
             # Add dx/dy to the limits to ensure the wireframe covers all of each pixel
             # as the xx and yy coordinates only give the centre of each pixel
-            dx = abs(xx[0][1] - xx[0][0])/2
+            dx = abs(xx[0][1] - xx[0][0]) / 2
             ax.set_xlim(np.nanmin(xx) - dx, np.nanmax(xx) + dx)
-            dy = abs(yy[1][0] - yy[0][0])/2
+            dy = abs(yy[1][0] - yy[0][0]) / 2
             ax.set_ylim(np.nanmin(yy) - dy, np.nanmax(yy) + dy)
 
         return self._get_wireframe_overlay(
