+++ planetmapper/observation.py
@@ -12,7 +12,8 @@ from astropy.io import fits
 from astropy.utils.exceptions import AstropyWarning
 
 from . import common, utils
-from .body_xy import BodyXY, Unpack, _cache_clearable_result, _MapKwargs
+from .base import _cache_clearable_result
+from .body_xy import BodyXY, Unpack, _MapKwargs
 from .progress import SaveMapProgressHookCLI, SaveNavProgressHookCLI, progress_decorator
 
 T = TypeVar('T')
